requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-allauth==0.38.0
django-appconf==1.0.2
django-bootstrap-pagination==1.7.0
django-bootstrap3==11.0.0
django-imagekit==4.0.2
django-import-export==1.1.0
django-seed==0.1.9
et-xmlfile==1.0.1
Faker==1.0.0
idna==2.7
jdcal==1.4
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.11
pilkit==2.0
Pillow==5.3.0
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
